+++ daemon/start.go
@@ -164,11 +164,12 @@ func (daemon *Daemon) containerStart(ctx context.Context, daemonCfg *configStore
 		return err
 	}
 
-	m, err := daemon.setupMounts(container)
+	m, cleanup, err := daemon.setupMounts(container)
 	if err != nil {
 		return err
 	}
 	mnts = append(mnts, m...)
+	defer cleanup()
 
 	spec, err := daemon.createSpec(ctx, daemonCfg, container, mnts)
 	if err != nil {